Extract fmtDate helpers to shared utils (per Parker's review)
Moved fmtDate, fmtDateFull, fmtDateInput from contact-block.tsx and
crm-page.tsx into lib/utils.ts to eliminate duplication.
